refactor: move wiggle onto Pipette
